Fix Clang compiler warnings

+++ src/p_inter.c
@@ -68,7 +68,7 @@ P_GiveAmmo
     if (ammo == am_noammo)
         return false;
 
-    if (ammo < 0 || ammo > NUMAMMO)
+    if ((int)ammo < 0 || (int)ammo > (int)NUMAMMO)
         I_Error ("P_GiveAmmo: bad type %i", ammo);
 
     if ( player->ammo[ammo] == player->maxammo[ammo]  )
